% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{somar}
\alias{somar}
\title{Random Effects SO MA of overlapping studies}
\usage{
somar(ES, SE, cor_mat, iter, type)
}
\arguments{
\item{ES}{Effect sizes from the meta analyses}

\item{SE}{Standard errors corresponding to the effect sizes}

\item{cor_mat}{Correlation matrix of the effect sizes}

\item{iter}{Number of iterations}

\item{type}{ML or REML}
}
\value{
* `Mean ES` mean effect size estimate

* `SE` standard error

* `Q` test statistic of test for heterogeneity

* `df` degrees of freedom of test for heterogeneity

* `p` p-value of test for heterogeneity

* `tau` measure of heterogeneity
}
\description{
Random Effects SO MA of overlapping studies
}
